+++ src/default-behavior.js
@@ -168,10 +168,35 @@ var createDefaultBehavior = function(editable) {
       log('Default move behavior');
     },
 
-    clipboard: function(element, action, cursor) {
-      if (action === 'paste') {
-        clipboard.paste(element, action, cursor, document);
+    paste: function(element, blocks, cursor) {
+      var fragment;
+
+      var firstBlock = blocks[0];
+      cursor.insertBefore(firstBlock);
+
+      if (blocks.length <= 1) {
+        cursor.setVisibleSelection();
+      } else {
+        var parent = element.parentNode;
+        var currentElement = element;
+
+        for (var i = 1; i < blocks.length; i++) {
+          var newElement = element.cloneNode(false);
+          if (newElement.id) newElement.removeAttribute('id');
+          fragment = content.createFragmentFromString(blocks[i]);
+          $(newElement).append(fragment);
+          parent.insertBefore(newElement, currentElement.nextSibling);
+          currentElement = newElement;
+        }
+
+        // focus last element
+        cursor = editable.createCursorAtEnd(currentElement);
+        cursor.setVisibleSelection();
       }
+    },
+
+    clipboard: function(element, action, cursor) {
+      log('Default clipboard behavior');
     }
   };
 };