+++ Changelog.md
@@ -1,3 +1,10 @@
+
+# v0.5.0
+
+- Split pasted content into blocks [#97](https://github.com/upfrontIO/editable.js/pull/97)
+- Adopt elements to iframe (set correct ownerDocument) [#98](https://github.com/upfrontIO/editable.js/pull/98)
+
+
 # v0.4.4
 
 #### Bugfixes